component folder structure
